Trait abomonation::size::AbomonationSize

pub trait AbomonationSize: Abomonation + Sized {
    fn extent(&self) -> usize { ... }

    fn measure(&self) -> usize { ... }
}

Sizing information for typed data in number of bytes required.a

Provided Methods

fn extent(&self) -> usize

Reports the number of further bytes required to entomb self.

fn measure(&self) -> usize

Reports the number of bytes required to encode self.